move may_wrap_dim function to a general header file
