+++ torch/csrc/inductor/aoti_standalone/utils.h
@@ -25,3 +25,15 @@ using AOTITorchError = int32_t;
     return AOTI_TORCH_FAILURE;                            \
   }                                                       \
   return AOTI_TORCH_SUCCESS;
+
+
+
+inline int64_t maybe_wrap_dim(int64_t dim, int64_t ndim) {
+  if (dim < 0) {
+    dim += ndim;
+  }
+  if (dim < 0 || dim >= ndim) {
+    throw std::runtime_error("dimension out of range");
+  }
+  return dim;
+}
